refactor(splunk): simplify data manager lifecycle
