--- modules/integrations/splunk_cloud_data_manager/data_input/scripts/splunk_data_manager/__init__.py
@@ -1,46 +1,0 @@
-"""Reusable Splunk Data Manager lifecycle package."""
-
-from . import cli
-from .cli import config_from_mapping, main
-from .client import SplunkWebClient, UrllibTransport
-from .input_state import (S3_DATASETS, input_uses_s3, s3_input_matches_request,
-                          s3_input_state, validate_input_document,
-                          wait_for_input)
-from .lifecycle import (DETAIL_RESPONSE_FIELDS, NOAH_TOKEN_PENDING,
-                        PUSH_HEC_CLEANUP_CATEGORIES, TOP_LEVEL_RESPONSE_FIELDS,
-                        build_delete_payload, create_integration,
-                        dataset_hec_categories, delete_integration,
-                        ensure_hec_tokens, get_integration,
-                        validate_cloudformation_template)
-from .runtime import (ArtifactPaths, HttpResponse, RuntimeConfig,
-                      SplunkHttpError, SplunkIntegrationError)
-
-__all__ = (
-    'ArtifactPaths',
-    'DETAIL_RESPONSE_FIELDS',
-    'HttpResponse',
-    'NOAH_TOKEN_PENDING',
-    'PUSH_HEC_CLEANUP_CATEGORIES',
-    'RuntimeConfig',
-    'S3_DATASETS',
-    'SplunkHttpError',
-    'SplunkIntegrationError',
-    'SplunkWebClient',
-    'TOP_LEVEL_RESPONSE_FIELDS',
-    'UrllibTransport',
-    'build_delete_payload',
-    'cli',
-    'config_from_mapping',
-    'create_integration',
-    'dataset_hec_categories',
-    'delete_integration',
-    'ensure_hec_tokens',
-    'get_integration',
-    'input_uses_s3',
-    'main',
-    's3_input_matches_request',
-    's3_input_state',
-    'validate_cloudformation_template',
-    'validate_input_document',
-    'wait_for_input',
-)
